Support color-incapable Hue devices; Use CIE XY colors
Fix #5. Fix #3. Note that color-temperature-only bulbs are currently
displayed as onOffLights since there is not yet a way to signal that the
only form of color control is temperature.

There may also be a mishap where an onOffLight is secretly an
onOffSwitch. This would require a modelid database which is something we
could provide but seems low priority for now.