Revert to HSV color space
The brightness of the CIE xyY color space suggested by the Philips Hue
API documentation appears to be incorrect. This may be a side effect of
not constraining the xy coordinates to the bulb's color gamut. Until a
more concerted effort can be made to figure out the root cause of the
brightness drops, this commit reverts back to using the HSV color space.